+++ src/app/store/entities/entities.actions.ts
@@ -2,15 +2,10 @@ import { createAction, props } from '@ngrx/store';
 import { EntityInterface } from './entities.interface';
 
 /**
- * Load Action will be used for NGRX Action emitter, when the Entities info Load Action is being called
+ * Fail Action will be used for NGRX Action emitter, when the Entities info is failed
  */
-export const Load = createAction('[ENTITIES] LOAD');
-
-/**
- * LoadFail Action will be used for NGRX Action emitter, when the Entities info is failed
- */
-export const LoadFail = createAction(
-	'[ENTITIES] LOAD FAIL',
+export const Fail = createAction(
+	'[ENTITIES] FAIL',
 	props<{
 		/**
 		 * errorMessage represents the errorMessage when type LOAD_FAIL is called
@@ -20,10 +15,34 @@ export const LoadFail = createAction(
 );
 
 /**
+ * Load Action will be used for NGRX Action emitter, when the Entities info Load Action is being called
+ */
+export const Load = createAction('[ENTITIES] LOAD');
+
+/**
  * LoadSuccess Action will be used for NGRX Action emitter, when the Entities info is succesfull
  */
 export const LoadSuccess = createAction(
 	'[ENTITIES] LOAD SUCCESS',
+	props<{
+		/**
+		 * entities represents the EntityInterface[] used for the reducer
+		 */
+		entities: EntityInterface[];
+	}>(),
+);
+
+/**
+ * Load Action will be used for NGRX Action emitter, when the Entities info Load Action is being called
+ */
+export const Map = createAction('[ENTITIES] MAP');
+
+/**
+ * MapSuccess Action will be used for NGRX Action emitter, when the Entities info is mapped succesfull
+ * NOTE: MapSuccess wont use the EntityMap functionality, due to the lack of testing in effects
+ */
+export const MapSuccess = createAction(
+	'[ENTITIES] MAP SUCCESS',
 	props<{
 		/**
 		 * entities represents the EntityInterface[] used for the reducer