+++ src/app/store/entities/entities.reducer.spec.ts
@@ -1,4 +1,4 @@
-import { initialState as mockStore } from '@testing/mock-store';
+import { cloneState, initialState as mockStore } from '@testing/mock-store';
 
 import * as entitiesActions from './entities.actions';
 import { initialState, reducer } from './entities.reducer';
@@ -19,7 +19,7 @@ describe('Reducers: Entities reducer', () => {
 
 	it('should not load entities', () => {
 		const payload = { errorMessage: 'global.something-went-wrong' };
-		const action = entitiesActions.LoadFail(payload);
+		const action = entitiesActions.Fail(payload);
 		const state = reducer(initialState, action);
 
 		expect(state.isLoading).toEqual(false);
@@ -46,5 +46,39 @@ describe('Reducers: Entities reducer', () => {
 		const state = reducer(initialState, action);
 
 		expect(state.errorMessage).toEqual(null);
+	});
+
+	describe('getDealsSuccess', () => {
+		let state = initialState;
+		const action = entitiesActions.MapSuccess({
+			entities: [
+				mockStore.entitiesState.entities[
+					'93dd0fd6-fd8c-4c70-a213-cb76d1ef6eda'
+				],
+			],
+		});
+		it('should add entities', () => {
+			state = reducer(initialState, action);
+			expect(state.entities).toEqual({
+				'93dd0fd6-fd8c-4c70-a213-cb76d1ef6eda':
+					mockStore.entitiesState.entities[
+						'93dd0fd6-fd8c-4c70-a213-cb76d1ef6eda'
+					],
+			});
+		});
+		it('should map entities', () => {
+			const clonedState = cloneState(mockStore);
+			state = reducer(clonedState.entitiesState, action);
+			expect(state.entities).toEqual({
+				'93dd0fd6-fd8c-4c70-a213-cb76d1ef6eda':
+					mockStore.entitiesState.entities[
+						'93dd0fd6-fd8c-4c70-a213-cb76d1ef6eda'
+					],
+			});
+		});
+		it('should set isLoading to false', () => {
+			state = reducer(initialState, action);
+			expect(state.isLoading).toEqual(false);
+		});
 	});
 });